Changes to the proto (double -> float) and gym-uds-server to adapt for instances that are not a single element. Sample transforms from nparray to python list.

+++ binding-cpp/include/gym-uds.h
@@ -12,9 +12,9 @@
 
 namespace gym_uds
 {
-using action_t = std::vector<double>;
-using observation_t = std::vector<double>;
-using state_t = std::tuple<observation_t, double, bool>;
+using action_t = std::vector<float>;
+using observation_t = std::vector<float>;
+using state_t = std::tuple<observation_t, float, bool>;
 
 
 class EnvironmentClient
